requirements.txt
streamlit>=1.28.0
pypdf
langchain
chromadb
pytest
boto3
langchain-community
langchain-ollama
langchain-text-splitters
